Use static_cast<> for size_t arguments
- fixes Warnings on Windows (where size_t is 64 bit)

+++ unitTestApp/src/RequestQueueBatcherTest.cpp
@@ -67,7 +67,7 @@ void
 TestDumper::processRequests(std::vector<std::shared_ptr<TestCargo>> &batch)
 {
     noOfBatches++;
-    batchSizes.push_back(batch.size());
+    batchSizes.push_back(static_cast<unsigned int>(batch.size()));
     std::vector<unsigned int> data;
     bool done = false;
